✨feat: character can only move in walkable area

+++ crates/world_generator/src/main.rs
@@ -27,15 +27,15 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 // 针对wall做一下特殊处理
                 if tileset_def.identifier == "Wall" {
                     let size = [
-                        tileset_def.tile_grid_size as u32,
-                        (tileset_def.tile_grid_size * 2) as u32,
+                        tileset_def.tile_grid_size as i32,
+                        (tileset_def.tile_grid_size * 2) as i32,
                     ];
                     tileset.tilt = true;
                     for y in (0..tileset_def.c_hei).step_by(2) {
                         for x in 0..tileset_def.c_wid {
                             let min = [
-                                (x as f32) * (tileset_def.tile_grid_size as f32),
-                                (y as f32) * (tileset_def.tile_grid_size as f32),
+                                (x as i32) * (tileset_def.tile_grid_size as i32),
+                                (y as i32) * (tileset_def.tile_grid_size as i32),
                             ];
                             tileset.tiles.insert(count, Rect { min, size });
                             count += 1;
@@ -43,14 +43,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                     }
                 } else {
                     let size = [
-                        tileset_def.tile_grid_size as u32,
-                        tileset_def.tile_grid_size as u32,
+                        tileset_def.tile_grid_size as i32,
+                        tileset_def.tile_grid_size as i32,
                     ];
                     for y in 0..tileset_def.c_hei {
                         for x in 0..tileset_def.c_wid {
                             let min = [
-                                (x as f32) * (tileset_def.tile_grid_size as f32),
-                                (y as f32) * (tileset_def.tile_grid_size as f32),
+                                (x as i32) * (tileset_def.tile_grid_size as i32),
+                                (y as i32) * (tileset_def.tile_grid_size as i32),
                             ];
                             tileset.tiles.insert(count, Rect { min, size });
                             count += 1;
@@ -83,8 +83,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             let mut rooms = vec![];
             for level in project.levels.iter() {
                 let mut room = RoomModel {
-                    world_pos: [level.world_x as f32, (-level.world_y) as f32],
-                    size: [level.px_wid as u32, level.px_hei as u32],
+                    world_pos: [level.world_x as i32, (-level.world_y) as i32],
+                    size: [level.px_wid as i32, level.px_hei as i32],
                     ..Default::default()
                 };
                 let mut walkable_area = vec![];
@@ -92,8 +92,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                     match layer.identifier.as_str() {
                         "Light" => {
                             for entity in layer.entity_instances.iter() {
-                                let x = entity.px.0 as f32;
-                                let y = -(entity.px.1 as f32);
+                                let x = entity.px.0 as i32;
+                                let y = -(entity.px.1 as i32);
                                 let mut color = None;
                                 let mut alpha = None;
                                 let mut inner = None;
@@ -110,7 +110,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                                 let inner = inner.unwrap();
                                 room.lights.push(Light {
                                     // TODO 这里的问题是如何让灯光低于roof但是能够让光扩散的足够开
-                                    pos: [x, y, if inner { 32.0 } else { 0.0 }],
+                                    pos: [x, y, if inner { 32 } else { 0 }],
                                     color: [color[0], color[1], color[2], alpha],
                                 })
                             }
@@ -128,14 +128,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                                 for tile in layer.grid_tiles.iter() {
                                     if let Some((index, rect)) =
                                         used_tileset.tiles.iter().find(|(_, rect)| {
-                                            rect.min[0] == tile.src.0 as f32
-                                                && rect.min[1] == tile.src.1 as f32
+                                            rect.min[0] == tile.src.0 as i32
+                                                && rect.min[1] == tile.src.1 as i32
                                         })
                                     {
                                         let width = rect.size[0];
                                         let height = rect.size[1];
-                                        let x = tile.px.0 as f32 + (width / 2) as f32;
-                                        let y = -(tile.px.1 as f32) - (height / 2) as f32;
+                                        let x = tile.px.0 as i32 + (width / 2) as i32;
+                                        let y = -(tile.px.1 as i32) - (height / 2) as i32;
                                         tile_group.tiles.push(Tile {
                                             pos: [x, y],
                                             index: *index,
@@ -143,7 +143,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                                         match layer.identifier.as_str() {
                                             "Carpet_Blue" | "Carpet_Red" | "Floor_Brick"
                                             | "Initial_Floor" => walkable_area.push(Rect {
-                                                min: [tile.px.0 as f32, -(tile.px.1 as f32)],
+                                                min: [tile.px.0 as i32, -(tile.px.1 as i32)],
                                                 size: [width, height],
                                             }),
                                             _ => {}
