+++ src/utils.rs
@@ -8,11 +8,11 @@ use crate::resource::SCALE_RATIO;
 pub fn tile_wall_sprite(
     mesh: &Handle<Mesh>,
     material: &Handle<StandardMaterial>,
-    pos: [f32; 2],
+    pos: [i32; 2],
     height: f32,
 ) -> PbrBundle {
-    let x = pos[0] * SCALE_RATIO;
-    let y = pos[1] * SCALE_RATIO;
+    let x = pos[0] as f32 * SCALE_RATIO;
+    let y = pos[1] as f32 * SCALE_RATIO;
     let z = (height * SCALE_RATIO / 3.0_f32.sqrt()) / 2.0;
     PbrBundle {
         mesh: mesh.clone(),
@@ -26,10 +26,10 @@ pub fn tile_wall_sprite(
 pub fn plane_pbr_sprite(
     mesh: &Handle<Mesh>,
     material: &Handle<StandardMaterial>,
-    pos: [f32; 2],
+    pos: [i32; 2],
 ) -> PbrBundle {
-    let x = pos[0] * SCALE_RATIO;
-    let y = pos[1] * SCALE_RATIO;
+    let x = pos[0] as f32 * SCALE_RATIO;
+    let y = pos[1] as f32 * SCALE_RATIO;
     let z = 0.0;
     PbrBundle {
         mesh: mesh.clone(),
@@ -39,10 +39,10 @@ pub fn plane_pbr_sprite(
     }
 }
 
-pub fn point_light(pos: [f32; 3], color: [u8; 4]) -> PointLightBundle {
-    let x = pos[0] * SCALE_RATIO;
-    let y = pos[1] * SCALE_RATIO;
-    let z = pos[2] * SCALE_RATIO;
+pub fn point_light(pos: [i32; 3], color: [u8; 4]) -> PointLightBundle {
+    let x = pos[0] as f32 * SCALE_RATIO;
+    let y = pos[1] as f32 * SCALE_RATIO;
+    let z = pos[2] as f32 * SCALE_RATIO;
     let intensity = if z > 0.0 { 5000.0 } else { 1000.0 };
     PointLightBundle {
         point_light: PointLight {
